requirements.txt
streamlit
supabase
google-generativeai
pydantic
PyPDF2
python-docx
pandas
plotly
python-dotenv
tiktoken
watchdog
